requirements.txt
blinker==1.5
click==8.0.3
colorama==0.4.6
dnspython==2.2.1
email-validator==1.3.0
Flask==2.0.2
Flask-Mail==0.9.1
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.23.5
paho-mqtt==1.6.1
pandas==1.5.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.6
six==1.16.0
Werkzeug==2.0.3
WTForms==3.0.1
